AMLS-4887: Simple code cleanup
