+++ app/controllers/businessdetails/RegisteredOfficeNonUKController.scala
@@ -54,7 +54,6 @@ class RegisteredOfficeNonUKController @Inject ()(
               registeredOffice <- businessDetails.registeredOffice
             } yield Form2[RegisteredOffice](registeredOffice)) getOrElse EmptyForm
             Ok(registered_office_non_uk(form, edit, autoCompleteService.getCountries))
-
         }
   }
